moved upload out of client
